+++ just_audio/android/src/main/java/com/ryanheise/just_audio/AudioPlayer.java
@@ -11,6 +11,7 @@ import android.os.Looper;
 import com.google.android.exoplayer2.C;
 import com.google.android.exoplayer2.DefaultLivePlaybackSpeedControl;
 import com.google.android.exoplayer2.DefaultLoadControl;
+import com.google.android.exoplayer2.DefaultRenderersFactory;
 import com.google.android.exoplayer2.ExoPlaybackException;
 import com.google.android.exoplayer2.LivePlaybackSpeedControl;
 import com.google.android.exoplayer2.LoadControl;
@@ -21,7 +22,7 @@ import com.google.android.exoplayer2.Player;
 import com.google.android.exoplayer2.Player.PositionInfo;
 import com.google.android.exoplayer2.ExoPlayer;
 import com.google.android.exoplayer2.Timeline;
-import com.google.android.exoplayer2.TracksInfo;
+import com.google.android.exoplayer2.Tracks;
 import com.google.android.exoplayer2.audio.AudioAttributes;
 import com.google.android.exoplayer2.extractor.DefaultExtractorsFactory;
 import com.google.android.exoplayer2.metadata.Metadata;
@@ -90,6 +91,7 @@ public class AudioPlayer implements MethodCallHandler, Player.Listener, Metadata
     private int errorCount;
     private AudioAttributes pendingAudioAttributes;
     private LoadControl loadControl;
+    private boolean offloadSchedulingEnabled;
     private LivePlaybackSpeedControl livePlaybackSpeedControl;
     private List<Object> rawAudioEffects;
     private List<AudioEffect> audioEffects = new ArrayList<AudioEffect>();
@@ -132,9 +134,10 @@ public class AudioPlayer implements MethodCallHandler, Player.Listener, Metadata
         }
     };
 
-    public AudioPlayer(final Context applicationContext, final BinaryMessenger messenger, final String id, Map<?, ?> audioLoadConfiguration, List<Object> rawAudioEffects) {
+    public AudioPlayer(final Context applicationContext, final BinaryMessenger messenger, final String id, Map<?, ?> audioLoadConfiguration, List<Object> rawAudioEffects, Boolean offloadSchedulingEnabled) {
         this.context = applicationContext;
         this.rawAudioEffects = rawAudioEffects;
+        this.offloadSchedulingEnabled = offloadSchedulingEnabled != null ? offloadSchedulingEnabled : false;
         methodChannel = new MethodChannel(messenger, "com.ryanheise.just_audio.methods." + id);
         methodChannel.setMethodCallHandler(this);
         eventChannel = new BetterEventChannel(messenger, "com.ryanheise.just_audio.events." + id);
@@ -217,9 +220,9 @@ public class AudioPlayer implements MethodCallHandler, Player.Listener, Metadata
     }
 
     @Override
-    public void onTracksInfoChanged(TracksInfo tracks) {
-        for (int i = 0; i < tracks.getTrackGroupInfos().size(); i++) {
-            TrackGroup trackGroup = tracks.getTrackGroupInfos().get(i).getTrackGroup();
+    public void onTracksChanged(Tracks tracks) {
+        for (int i = 0; i < tracks.getGroups().size(); i++) {
+            TrackGroup trackGroup = tracks.getGroups().get(i).getMediaTrackGroup();
 
             for (int j = 0; j < trackGroup.length; j++) {
                 Metadata metadata = trackGroup.getFormat(j).metadata;
@@ -727,7 +730,11 @@ public class AudioPlayer implements MethodCallHandler, Player.Listener, Metadata
             if (livePlaybackSpeedControl != null) {
                 builder.setLivePlaybackSpeedControl(livePlaybackSpeedControl);
             }
+            if (offloadSchedulingEnabled) {
+                builder.setRenderersFactory(new DefaultRenderersFactory(context).setEnableAudioOffload(true));
+            }
             player = builder.build();
+            player.experimentalSetOffloadSchedulingEnabled(offloadSchedulingEnabled);
             setAudioSessionId(player.getAudioSessionId());
             player.addListener(this);
         }